feat: primeira revisao de repositories

+++ refoods/src/main/java/com/projeto/ReFood/repository/CardRepository.java
@@ -3,4 +3,4 @@ package com.projeto.ReFood.repository;
 import com.projeto.ReFood.model.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Integer> {}
+public interface CardRepository extends JpaRepository<Card, Long> {}
