+++ refoods/src/main/java/com/projeto/ReFood/repository/FavoriteRepository.java
@@ -3,5 +3,5 @@ package com.projeto.ReFood.repository;
 import com.projeto.ReFood.model.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
 }
