+++ refoods/src/main/java/com/projeto/ReFood/repository/NotificationRepository.java
@@ -2,6 +2,10 @@ package com.projeto.ReFood.repository;
 
 import com.projeto.ReFood.model.Notification;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {}
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
+
+}
