+++ refoods/src/main/java/com/projeto/ReFood/repository/OrderRepository.java
@@ -3,4 +3,4 @@ package com.projeto.ReFood.repository;
 import com.projeto.ReFood.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {}
+public interface OrderRepository extends JpaRepository<Order, Long> {}
