+++ refoods/src/main/java/com/projeto/ReFood/repository/ProductRepository.java
@@ -3,4 +3,4 @@ package com.projeto.ReFood.repository;
 import com.projeto.ReFood.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {}
+public interface ProductRepository extends JpaRepository<Product, Long> {}
